add security to products and profile views

+++ products/views.py
@@ -1,11 +1,12 @@
 from django.shortcuts import render, redirect, reverse, get_object_or_404
 from django.contrib import messages
+from django.contrib.auth.decorators import login_required
 from django.db.models.functions import Lower
-from django.db.models import Q # this is django built in and generates a search query
+from django.db.models import Q
 from .models import Product, Category, Manufacturer
 from .forms import ProductForm
 
-# Create your views here.
+
 def all_products(request):
     """ 
     A view to show all products,
@@ -92,10 +93,15 @@ def product_detail(request, product_id):
     return render(request, 'products/product_detail.html', context)
 
 
+@login_required
 def add_product(request):
   """ 
   Add a product to the store
   """
+  if not request.user.is_superuser:
+    messages.error(request, 'Only staff have access to this feature.')
+    return redirect(reverse('home'))
+
   if request.method == 'POST':
     form = ProductForm(request.POST, request.FILES)
     if form.is_valid():
@@ -114,10 +120,15 @@ def add_product(request):
   return render(request, template, context)
 
 
+@login_required
 def edit_product(request, product_id):
   """ 
   Edit a product in the store
   """
+  if not request.user.is_superuser:
+    messages.error(request, 'Only staff have access to this feature.')
+    return redirect(reverse('home'))
+
   product = get_object_or_404(Product, pk=product_id)  
 
   if request.method == 'POST':
@@ -141,10 +152,16 @@ def edit_product(request, product_id):
   }
   return render(request, template, context)
   
+
+@login_required
 def delete_product(request, product_id):
   """
   Delete a product from the store
   """
+  if not request.user.is_superuser:
+    messages.error(request, 'Only staff have access to this feature.')
+    return redirect(reverse('home'))
+
   product = get_object_or_404(Product, pk=product_id)
   product.delete()
   messages.success(request, 'Your selected product has been removed from the store')
